+++ docs/src/news.md
@@ -3,6 +3,7 @@
 ## *Version 0.9.6*
 
 - News:
+    - Function **expand_grid**: Build a dataframe with all the combinations of the entered parameter values.
     - Fonctions **sampbag**, **baggr**: bagging regression.
     - Fonction **protoplsr**: Averaging PLSR models built on the neighborhood of prototype observations.
     - Function **rclustplsr**: Random clustered PLSR.